Solved issue #19
Issue: The hot dog and burger orders would not call the functions in order.js
Problem: The input element for the size was required but for these 2 orders the size is not listed. In the orderAll() function the try statement is 'breaked' when a required element is not filled in.
Resolution: Removed attribute 'required' from size input element